Renamed `InternalError.ts` to `internalerror.ts`.
